+++ mongo/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## To be Released
 
+## v2.0.0
+
 * feat(document): `Validate` returns an `error` [BREAKING CHANGE]
 * feat(document): remove unused `ValidateWithInternalError` [BREAKING CHANGE]